// src/fileStorageApi/v3/models/file-record-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software File Storage API
 * This is the API documentation for the Schulcloud-Verbund-Software File Storage API
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FileRecordResponse } from './file-record-response';

/**
 * 
 * @export
 * @interface FileRecordListResponse
 */
export interface FileRecordListResponse {
    /**
     * The items for the current page.
     * @type {Array<FileRecordResponse>}
     * @memberof FileRecordListResponse
     */
    data: Array<FileRecordResponse>;
    /**
     * The total amount of items.
     * @type {number}
     * @memberof FileRecordListResponse
     */
    total: number;
    /**
     * The amount of items skipped from the start.
     * @type {number}
     * @memberof FileRecordListResponse
     */
    skip: number;
    /**
     * The page size of the response.
     * @type {number}
     * @memberof FileRecordListResponse
     */
    limit: number;
}