// src/fileStorageApi/v3/models/file-record-parent-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */
export enum FileRecordParentType {
    USERS = 'users',
    SCHOOLS = 'schools',
    COURSES = 'courses',
    TASKS = 'tasks',
    LESSONS = 'lessons',
    SUBMISSIONS = 'submissions',
    BOARDNODES = 'boardnodes'
}